Use BigInteger for BencodedInt
The bencoding specification defines BencodedInt as any base 10 ASCII text
representation of an integer. This can be arbitrarily large, and should not
be limited to the storage capacities of `Int` or `Long`.

Therefore, we include a multiplatform big number library.

Closes #1

+++ corbit-bencoding/src/commonMain/kotlin/com/corbit/bencoding/BencodedData.kt
@@ -1,12 +1,13 @@
 package com.corbit.bencoding
 
 import com.corbit.binary.BinaryString
+import com.ionspin.kotlin.bignum.integer.BigInteger
 
 sealed class BencodedData {
     fun encode(): BinaryString = Encoder(this).encode()
 }
 
-data class BencodedInt(val value: Long) : BencodedData()
+data class BencodedInt(val value: BigInteger) : BencodedData()
 
 data class BencodedString(val value: BinaryString) : BencodedData()
