+++ corbit-bencoding/src/commonTest/kotlin/com/corbit/bencoding/EncoderTest.kt
@@ -2,6 +2,7 @@ package com.corbit.bencoding
 
 import com.corbit.binary.BinaryString
 import com.corbit.binary.asBinaryString
+import com.ionspin.kotlin.bignum.integer.toBigInteger
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -9,7 +10,7 @@ class EncoderTest {
 
     @Test fun `bencoded int can properly be encoded to binary data`() {
         // arrange
-        val data = BencodedInt(123)
+        val data = BencodedInt(123.toBigInteger())
 
         // act
         val result = data.encode()
@@ -66,7 +67,7 @@ class EncoderTest {
     @Test fun `trivial bencoded list is encoded to proper binary data`() {
         // arrange
         val data = BencodedList(listOf(
-            BencodedInt(123),
+            BencodedInt(123.toBigInteger()),
             BencodedString("eggs".asBinaryString())
         ))
 
@@ -84,7 +85,7 @@ class EncoderTest {
         // arrange
         val data = BencodedList(listOf(
             BencodedList(listOf(
-                BencodedInt(123),
+                BencodedInt(123.toBigInteger()),
                 BencodedString("eggs".asBinaryString())
             ))
         ))
@@ -116,7 +117,7 @@ class EncoderTest {
     @Test fun `trivial bencoded dictionary is encoded to proper binary data`() {
         // arrange
         val data = BencodedDict(mapOf(
-            BencodedString("foo".asBinaryString()) to BencodedInt(123),
+            BencodedString("foo".asBinaryString()) to BencodedInt(123.toBigInteger()),
             BencodedString("bar".asBinaryString()) to BencodedString("baz".asBinaryString())
         ))
 
@@ -134,7 +135,7 @@ class EncoderTest {
         // arrange
         val data = BencodedDict(mapOf(
             BencodedString("map".asBinaryString()) to BencodedDict(mapOf(
-                BencodedString("foo".asBinaryString()) to BencodedInt(123),
+                BencodedString("foo".asBinaryString()) to BencodedInt(123.toBigInteger()),
                 BencodedString("bar".asBinaryString()) to BencodedString("baz".asBinaryString())
             ))
         ))
